removed unused components in admin dashboard
