+++ src/App.js
@@ -28,7 +28,6 @@ import MentorTable from "./components/dashboard/admin-dashboard/MentorTable";
 import TeacherTable from "./components/dashboard/admin-dashboard/TeacherTable";
 import mentorDashboard from "./components/dashboard/mentor-dashboard/mentorDashboard.js";
 import TeacherDashboard from "./components/dashboard/teacher-dashboard/TeacherDashboard";
-import Dashboard from "./components/dashboard/admin-dashboard/Dashboard.js";
 import DashboardTeacher from "./components/dashboard/teacher-dashboard/DashboardTeacher";
 
 import "./App.css";
@@ -124,12 +123,10 @@ class App extends React.Component {
             <Route exact path="/" component={LandingPage} />
             <Route path="/mentors" component={MentorList} />
             <Route path="/FAQ" component={FAQ} />
-            {/* //! testing new dashboard */}
-            <Route path="/admin-dashboard" component={Dashboard} />
+            <Route path="/admin-dashboard" component={AdminDashboard} />
             <Route path="/approved-teachers" component={ApprovedTeacherList} />
             <Route path="/approved-mentors" component={ApprovedMentorList} />
             <Route path="/view-profile" component={ViewUserProfile} />
-            <Route path="/test-dashboard" component={Dashboard} />
             <Route
               exact
               path="/signup"
