+++ src/components/dashboard/admin-dashboard/AdminDashboard.js
@@ -1,107 +1,230 @@
-// import React, { Component } from "react";
-// import { MDBContainer, MDBRow, MDBCol } from "mdbreact";
-// import { Link } from "react-router-dom";
-// import { connect } from "react-redux";
-// import Button from "@material-ui/core/Button";
-// import { firestore } from "../../../config/fbConfig.js";
-// import MentorTable from "./MentorTable.js";
-// import TeacherTable from "./TeacherTable.js";
-// import ApprovedMentorList from "../ApprovedMentorList.js";
-// import SideBar from "../SideBar";
-// // import { QuerySnapshot } from "@google-cloud/firestore";
+import React, { useState, useEffect } from "react";
+import { withRouter } from "react-router";
+import { firestore } from "../../../config/fbConfig.js";
+import { connect } from "react-redux";
+import MentorTable from "./MentorTable.js";
+import TeacherTable from "./TeacherTable.js";
+//styles
+import clsx from "clsx";
+import { makeStyles } from "@material-ui/core/styles";
+import CssBaseline from "@material-ui/core/CssBaseline";
+import Drawer from "@material-ui/core/Drawer";
+import AppBar from "@material-ui/core/AppBar";
+import Toolbar from "@material-ui/core/Toolbar";
+import List from "@material-ui/core/List";
+import Typography from "@material-ui/core/Typography";
+import Divider from "@material-ui/core/Divider";
+import IconButton from "@material-ui/core/IconButton";
+import Badge from "@material-ui/core/Badge";
+import Container from "@material-ui/core/Container";
+import Grid from "@material-ui/core/Grid";
+import Paper from "@material-ui/core/Paper";
+import Link from "@material-ui/core/Link";
+import MenuIcon from "@material-ui/icons/Menu";
+import ChevronLeftIcon from "@material-ui/icons/ChevronLeft";
+import NotificationsIcon from "@material-ui/icons/Notifications";
+import { mainListItems, secondaryListItems } from "./listItems";
 
-// import "../Dashboard.css";
+const drawerWidth = 240;
 
-// class AdminDashboard extends Component {
-//   state = {
-//     users: [],
-//     userType: ""
-//   };
+const useStyles = makeStyles(theme => ({
+  root: {
+    display: "flex"
+  },
+  toolbar: {
+    paddingRight: 24 // keep right padding when drawer closed
+  },
+  toolbarIcon: {
+    display: "flex",
+    alignItems: "center",
+    justifyContent: "flex-end",
+    padding: "0 8px",
+    ...theme.mixins.toolbar
+  },
+  appBar: {
+    zIndex: theme.zIndex.drawer + 1,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    })
+  },
+  appBarShift: {
+    marginLeft: drawerWidth,
+    width: `calc(100% - ${drawerWidth}px)`,
+    transition: theme.transitions.create(["width", "margin"], {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  menuButton: {
+    marginRight: 36
+  },
+  menuButtonHidden: {
+    display: "none"
+  },
+  title: {
+    flexGrow: 1
+  },
+  drawerPaper: {
+    position: "relative",
+    whiteSpace: "nowrap",
+    width: drawerWidth,
+    transition: theme.transitions.create("width", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.enteringScreen
+    })
+  },
+  drawerPaperClose: {
+    overflowX: "hidden",
+    transition: theme.transitions.create("width", {
+      easing: theme.transitions.easing.sharp,
+      duration: theme.transitions.duration.leavingScreen
+    }),
+    width: theme.spacing(7),
+    [theme.breakpoints.up("sm")]: {
+      width: theme.spacing(9)
+    }
+  },
+  appBarSpacer: theme.mixins.toolbar,
+  content: {
+    flexGrow: 1,
+    height: "100vh",
+    overflow: "auto"
+  },
+  container: {
+    paddingTop: theme.spacing(4),
+    paddingBottom: theme.spacing(4)
+  },
+  paper: {
+    padding: theme.spacing(2),
+    display: "flex",
+    overflow: "auto",
+    flexDirection: "column"
+  },
+  fixedHeight: {
+    height: 240
+  }
+}));
 
-//   componentDidMount = async () => {
-//     let userArray = [];
-//     const userRef = firestore.collection("users");
-//     const userList = await userRef.get().then(querySnapshot => {
-//       querySnapshot.forEach(doc => {
-//         // console.log(doc.data());
-//         userArray.push({
-//           approval: doc.data().awaitingApproval,
-//           name: doc.data().fullName,
-//           userType: doc.data().userType,
-//           city: doc.data().city,
-//           stateProvince: doc.data().stateProvince,
-//           uid: doc.data().uid
-//         });
-//       });
-//     });
-//     this.setState({
-//       users: userArray
-//     });
-//     //! This is the fix for the refreshing as an admin and not being able to get back to this component.
-//     //! This also fixes accounts that are not admins being able to access the admin-dash
-//     //! Only problem is, when you are a non-admin account and attempt access, it flashes the admin dash before redirecting.
-//     setTimeout(() => {
-//       if (this.props.userInfo.userType !== "admin") {
-//         this.props.history.push("/");
-//       }
-//     }, 0);
-//   };
-//   render() {
-//     // const { auth, userInfo } = this.props;
-//     // console.log("auth", auth);
-//     // console.log("userinfo", userInfo);
+const AdminDashboard = props => {
+  console.log(props);
+  const [users, setUsers] = useState([]);
 
-//     return (
-//       <div className="dashboardContainer">
-//         <MDBContainer>
-//           <MDBRow id="dashboard-MDBRow">
-//             <div className="row justify-content-start">
-//               <MDBCol size="3" className="dashboard-sidemenu">
-//                 <div className="sidebar-info">
-//                   <img src="#" alt="profile photo" />
-//                   <h1>Norman Hill</h1>
-//                   <h3>Administrator</h3>
-//                 </div>
-//                 <div className="dashboard-sidemenu-btns">
-//                   <Button href="#">Schedule a Meeting</Button>
-//                   <Button
-//                     onClick={() =>
-//                       this.props.history.push("/approved-teachers")
-//                     }
-//                   >
-//                     View Approved Teachers
-//                   </Button>
-//                   <Button
-//                     onClick={() => this.props.history.push("/approved-mentors")}
-//                   >
-//                     View Approved Mentors
-//                   </Button>
-//                   <Button href="#">Start a Conversation</Button>
-//                 </div>
-//               </MDBCol>
-//             </div>
-//             <MDBCol size="9">
-//               <MentorTable
-//                 users={this.state.users}
-//                 history={this.props.history}
-//               />
-//               <TeacherTable
-//                 users={this.state.users}
-//                 history={this.props.history}
-//               />
-//             </MDBCol>
-//           </MDBRow>
-//         </MDBContainer>
-//       </div>
-//     );
-//   }
-// }
-// const mapStateToProps = state => {
-//   // console.log(state);
-//   return {
-//     auth: state.firebase.auth,
-//     users: state.firestore.ordered.users,
-//     userInfo: state.firebase.profile //need access to the users collection instead to check userType and render props in the tables
-//   };
-// };
-// export default connect(mapStateToProps)(AdminDashboard);
+  useEffect(() => {
+    dashboardAdmin();
+  }, []);
+
+  const dashboardAdmin = async () => {
+    let userArray = [];
+    const userRef = firestore.collection("users");
+    await userRef.get().then(querySnapshot => {
+      querySnapshot.forEach(doc => {
+        // console.log(doc.data());
+        userArray.push({
+          approval: doc.data().awaitingApproval,
+          name: doc.data().fullName,
+          userType: doc.data().userType,
+          city: doc.data().city,
+          stateProvince: doc.data().stateProvince,
+          uid: doc.data().uid
+        });
+      });
+    });
+    setUsers(userArray);
+    console.log(users);
+  };
+
+  const classes = useStyles();
+  const [open, setOpen] = useState(true);
+  const handleDrawerOpen = () => {
+    setOpen(true);
+  };
+  const handleDrawerClose = () => {
+    setOpen(false);
+  };
+
+  return (
+    <div className={classes.root}>
+      <CssBaseline />
+      <AppBar
+        position="absolute"
+        className={clsx(classes.appBar, open && classes.appBarShift)}
+      >
+        <Toolbar className={classes.toolbar}>
+          <IconButton
+            edge="start"
+            color="inherit"
+            aria-label="open drawer"
+            onClick={handleDrawerOpen}
+            className={clsx(
+              classes.menuButton,
+              open && classes.menuButtonHidden
+            )}
+          >
+            <MenuIcon />
+          </IconButton>
+          <Typography
+            component="h1"
+            variant="h6"
+            color="inherit"
+            noWrap
+            className={classes.title}
+          >
+            Welcome to your Dashboard
+          </Typography>
+          <IconButton color="inherit">
+            <Badge badgeContent={4} color="secondary">
+              <NotificationsIcon />
+            </Badge>
+          </IconButton>
+        </Toolbar>
+      </AppBar>
+      <Drawer
+        variant="permanent"
+        classes={{
+          paper: clsx(classes.drawerPaper, !open && classes.drawerPaperClose)
+        }}
+        open={open}
+      >
+        <div className={classes.toolbarIcon}>
+          <IconButton onClick={handleDrawerClose}>
+            <ChevronLeftIcon />
+          </IconButton>
+        </div>
+        <Divider />
+        <List>{mainListItems}</List>
+        <Divider />
+        <List>{secondaryListItems}</List>
+      </Drawer>
+      <main className={classes.content}>
+        <div className={classes.appBarSpacer} />
+        <Container maxWidth="lg" className={classes.container}>
+          <Grid container spacing={3}>
+            {/* Mentor Table */}
+            <Grid item xs={12}>
+              <Paper className={classes.paper}>
+                <MentorTable users={users} history={props.history} />
+              </Paper>
+            </Grid>
+            {/* Teacher Table */}
+            <Grid item xs={12}>
+              <Paper className={classes.paper}>
+                <TeacherTable users={users} history={props.history} />
+              </Paper>
+            </Grid>
+          </Grid>
+        </Container>
+      </main>
+    </div>
+  );
+};
+
+const mapStateToProps = state => {
+  // console.log(state);
+  return {
+    auth: state.firebase.auth,
+    users: state.firestore.ordered.users,
+    userInfo: state.firebase.profile //need access to the users collection instead to check userType and render props in the tables
+  };
+};
+export default withRouter(connect(mapStateToProps)(AdminDashboard));
